+++ neumorphism/itwillrock_mobile_ui/example/lib/navigation_service.dart
@@ -49,7 +49,7 @@ class NavigationService with ChangeNotifier {
     if (_navigationStack.length > 1) {
       _navigationStack.removeLast();
       _titleStack.removeLast();
-      _notify();
+      notifyListeners();
     }
   }